Add `group` text
Couldn't figure out how to add it to the individual song info in the wheel, so I used the general `ScreenSelectMusic_decorations/default.lua` file.

Couldn't resize & move the song's image either (not in this [or other] file) so I couldn't add it to that stack easily without nudging every single other element.

Sooo I decided to rotate it and throw it to the right of the wheel.